--- src/main/java/shop/tryit/domain/member/MemberFormDto.java
@@ -1,38 +1,0 @@
-package shop.tryit.domain.member;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotEmpty;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import shop.tryit.domain.common.Address;
-
-@Getter
-@NoArgsConstructor
-public class MemberFormDto {
-
-    @NotBlank(message = "이름은 필수 입력 값입니다.")
-    private String name;
-
-    @NotBlank(message = "이메일은 필수 입력 값입니다.")
-    @Email(message = "이메일 형식으로 입력해주세요.")
-    private String email;
-
-    @NotEmpty(message = "비밀번호는 필수 입력 값입니다.")
-    private String password;
-
-    @NotEmpty(message = "주소는 필수 입력 값입니다.")
-    private Address address;
-
-    private String phone;
-
-    @Builder
-    public MemberFormDto(String name, String email, String password, Address address, String phone) {
-        this.name = name;
-        this.email = email;
-        this.password = password;
-        this.address = address;
-        this.phone = phone;
-    }
-}
